Introduction of data-{country,state,enable,disable} attributes:
A peak's country and state are now specified by the data-country and
data-state attributes in the HTML. These are processed by the
JavaScript to determine whether a peak is in the U.S. and whether it's
in California. Previously this was done less elegantly and less
accurately by the JavaScript functions isUSPeak and isCAPeak which
could be defined differently for each peak list.

The default country is "US", and the default state is "CA". These can
be overridden by setting data-country and data-state on the first row
(TR element) of the peak table. These attributes can also be set on
the section header row to set the default for all peaks in a section,
and they can be set on an individual peak's row to apply to that peak.

Similarly, the data-enable and data-disable attributes can be used to
set/unset additional peak flags that can be referenced by the
JavaScript. Currently, there is only one flag that should be enabled
or disabled with these attributes: the "CC" flag determines whether to
include a link to ClosedContour.com in the geo-links popup (aka the
map link box). The "CC" flag is unset by default and should be set
only for Sierra peaks.