Two changes:
* The checkbox that toggles outline display now also toggles whether the query
for the outline geometry is issued. By default the checkbox is unchecked, so the
outline geometry will not be loaded. Checking the checkbox will load and display
the outline for the currently clicked point, and, as long as it's checked, the
outline will be loaded and displayed for future clicks (if the initial query by
lat/long returns a result for that overlay).

* The lat/long are now displayed at the top of the popup, and when clicked, the
latitude comma longitude will be copied to the clipboard.