Various changes. Most notably:
* Use rmgsc-haws1 instead of wildfire.cr.usgs.gov for GeoMAC layers. The latter
seems to load balance between rmgsc-haws1 and rmgsc-haws2 which seems to cause some
inconsistency problems across multiple queries since one query may go to haws1
while the next goes to haws2. The same polygon may have different ArcGIS object IDs
on the different servers. So querying for the geometry for an object ID returned
by a previous query may not return the desired geometry. It may not return any
geometry, or it may return the geometry for a totally different polygon. Also, the
VIIRS tiles don't seem to be in sync between the two rmgsc-haws servers.

* Added default minZoom and maxZoom to the options object used when creating a new
L.TileLayer or L.GridLayer. This prevents the numerous "Error in parsing value for
'z-index'" messages in the Web/JavaScript console. I think this is because the
_updateLevels function in Leaflet's src/layer/tile/GridLayer.js sets zIndex for the
tile container (div) element to maxZoom (or maxZoom minus a value) without checking
whether maxZoom is actually defined. So it could be setting it to undefined or NaN.

* Geometry queries can now be enabled by specifying the "qg" parameter in the query
string. The default outline color can also be specified. For example:
"qg=geomac_lp:ff00ff" will enable the geometry query for GeoMAC Latest Perimeters,
and the outline color will be set to magenta.

* Display the perimeter date/time when quering GeoMAC Latest Perimeters.

* Changed the "cc" query string flag to "clk". The GeoMAC map servers seem to block
requests when the HTTP Referer contains "&cc"!?

* The code for adding overlays, baselayers, and tile overlays to the layer control is
now a bit cleaner from the caller's perspective. The caller (pmap.html) now just has
to pass the root LCD (layer control definition) object once along with the paths to be
enabled (specified in the query string), and the code in pmap-lc.js will issue the
recursive call to the appropriate _add function and then enable the specified paths.